QC plot voor digital droplet pcr analyse

QC plot voor digital droplet pcr analyse

ddpcr_qc_calc(
  data,
  dilution_base = 20,
  dilution_factor = 5,
  blank = "blanco",
  limit_type = "theo",
  lod_min_positives = 0.05,
  lod_manual = NULL,
  loq_slope_tolerance = 0.6,
  loq_manual = NULL,
  log_add = 1e-04,
  ...
)

Arguments

data

toestelfile die op zijn minst Sample, Concentration en Threshold bevat, waarbij de samples alfabetisch staan volgens stijgende verdunning en er een blanco staal is. Ieder staal moet minstens enkele observaties bevatten

dilution_base

hoeveel keer is het eerste staal verdund. Standaard 20

dilution_factor

hoeveel keer wordt er extra verdund van staal tot staal. Standaard 5

blank

de naam van de Sample die als blanco moet gerekend worden

limit_type

welke waarde moet gebruikt worden om de LOD en LOQ op de grafiek te zetten. De theoretische concenteratie ("theo") of de gemeten concentratie ("meas")

lod_min_positives

minimum fractie van detecties om boven de LOD te komen. Standaard 1/20

lod_manual

een manuele LOD als je die zelf wil berekenen

loq_slope_tolerance

Bij weinig verdunning verwacht je de helling dichtbij 1, maar hoe meer verdunt hoe lager deze hellling. Deze parameter bevat de limiet voor wat nog als lineair bereik gerekend wordt. Standaard is dit op 0.60 gezet.

loq_manual

manuele waarde voor LOQ als je die zelf wil berekenen

log_add

aangezien er logtransformaties gebruikt worden, en een 0 niet kan omgezet worden, bevat deze parameter wat er bijgeteld wordt voor alle concentraties. Standaard 0.0001

...

extra parameters die doorgegeven worden aan de functie ddpcr_calculate_fit waarvan in deze functie gebruikt wordt gemaakt

Value

een ggplot object met de kwaliteitsfiguur of indien plot_it = FALSE een data.frame met alle nodige plotgegevens

Examples

if (FALSE) {
library(ggplot2)
fpath <- system.file("extdata", "ddPCR_voorbeelddata.csv",
  package = "inbolims"
)
example_data <- read.csv2(fpath, stringsAsFactors = FALSE)
plotdata <- ddpcr_qc_calc(example_data, estim_formula = "connect")
ggplot(plotdata)
summary(plotdata)
}