Get report config information

Get report config information

get_report_config_info(template = "default")

Arguments

template

a character vector of templates you want info from. If "all" then all templates are shown

Value

dataset wih report template information

Examples

{
get_report_config_info(template = "default")
}
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)
#> Rows: 65 Columns: 7
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (6): Type, Veldnaam, Tabel, Afkorting, Kolom, Beschrijving
#> dbl (1): template:default
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 65 × 7
#>    Type   Veldnaam         Tabel Afkorting Kolom Beschrijving `template:default`
#>    <chr>  <chr>            <chr> <chr>     <chr> <chr>                     <dbl>
#>  1 select ContractID       dimC… c         Cont… Contractnaam                  1
#>  2 select Klant            fact… r         Cust… Identificat…                  2
#>  3 select Project          fact… r         Proj… Projectnaam                   3
#>  4 select Verantwoordelij… dimP… p         Resp… Labo verant…                  4
#>  5 select LaboLimsGroep    dimP… p         Proj… Veiligheids…                  0
#>  6 select LimsStaalNummer  dimS… s         LIMS… VERPLICHT B…                  6
#>  7 select ExternSampleID   dimS… s         Fiel… Veldcode vo…                  7
#>  8 select LaboCode         dimS… s         LabS… De interne …                  8
#>  9 select OrigineelStaal   dimS… s         LIMS… Belangrijk:…                  9
#> 10 select SampleProduct    dimS… s         Prod… Product zoa…                 10
#> # ℹ 55 more rows