GGplot of ddPCR data (generated by ddPCR_qcalc)

GGplot of ddPCR data (generated by ddPCR_qcalc)

# S3 method for ddPCR
ggplot(
  data,
  mapping = NULL,
  colors = c("black", "red", "blue", "green4"),
  shapes = c(-1, 1, 4),
  linetypes = c("solid", "blank", "solid", "dashed"),
  lod_label = "LOD",
  loq_label = "LOQ",
  xlab = expression(paste(log[10], "(calculated DNA concentration)")),
  ylab = expression(paste(log[10], "(measured DNA concentration)")),
  log_add = 1e-04,
  ...
)

Arguments

data

dataset generated from ddPCR_qccalc. Must have at least these columns:

log10Concentration

log10Theoretical

which_data groeperingsvariabele bestaande uit 3 niveau's. De naam is vrij te kiezen, maar de alfabetische volgorde zal bepalen welke kleur en lijntype eraan gematcht worden

log10loq

log10lod

mapping

wordt niet gebruikt in dit geval, maar is nodig om de ggplot functie te kunnen gebruiken

colors

vector van lengte 4 met de kleuren van de verschillende plotdelen die bepaald worden door de volgorde van de niveau's in which_data

kleur 1 is het kleur voor de eerste groep (standaard de gefitte lijn)

kleur 2 is het kleur voor de tweede groep (standaard de observaties)

kleur 3 is het kleur voor de derde groep (standaard de theoretische )

kleur 4 is het kleur voor de LOD en LOQ lijnen

shapes

numerieke vector van lengte 3 die de symbolen overeenkomstig de groepen in which_data definieert. Bij een negatief getal wordt het punt niet getoond, symbool met nummer 1 is een lege bol, symbool met nummer 4 is een kruisje. Standaard wordt de eerste shape negatief gezet, omdat dit standaard overeenkomt met de gefitte lijn, waarvoor we de punten niet willen zien

linetypes

vector in dezelfde volgorde als colors die de lijntypes definieert. Voor een lijn die niet getekend moet worden geef je de waarde "blank", een volle lijn is "solid", een stippelliljn "dashed" en dan heb je nog heel wat andere types die je kan terugvinden via google "ggplot linetypes"

lod_label

tekst die bij de LOD lijn moet gezet worden

loq_label

tekst die bij de LOQ lijn moet gezet worden

xlab

astitel voor de x-as

ylab

astitel voor de y-as

log_add

hoeveel wordt bij de concentratie geteld om de log te nemen (log10(variabele + log_add)). Deze waarde wordt gebruikt om enkel de datarijen te selecteren met een concentratie groter dan 0, en moet overeenstemmen met de log_add in ddPCR_qc_calc

...

optional arguments (not used in this function, but needed for R CMD CHECK)

Value

ggplot object met 5 lagen: de geobserveerde waarden, de fit door de niet-nulwaarden van de geobserveerde waarden, de theoretische waarden, een vertikale lijn voor LOD, een vertikale lijn voor LOQ